# packages/api-schema/open-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/calendar/feeds": {
      "get": {
        "operationId": "CalendarController_getFeeds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array"
                  ],
                  "items": {
                    "type": [
                      "object"
                    ],
                    "properties": {
                      "id": {
                        "type": [
                          "string"
                        ]
                      },
                      "personName": {
                        "type": [
                          "string"
                        ]
                      },
                      "color": {
                        "type": [
                          "string"
                        ]
                      },
                      "icsUrl": {
                        "type": [
                          "string"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "personName",
                      "color",
                      "icsUrl"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Calendar"
        ]
      },
      "post": {
        "operationId": "CalendarController_addFeed",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": [
                  "object"
                ],
                "properties": {
                  "personName": {
                    "type": [
                      "string"
                    ]
                  },
                  "color": {
                    "type": [
                      "string"
                    ]
                  },
                  "icsUrl": {
                    "type": [
                      "string"
                    ]
                  }
                },
                "required": [
                  "personName",
                  "color",
                  "icsUrl"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "object"
                  ],
                  "properties": {
                    "id": {
                      "type": [
                        "string"
                      ]
                    },
                    "personName": {
                      "type": [
                        "string"
                      ]
                    },
                    "color": {
                      "type": [
                        "string"
                      ]
                    },
                    "icsUrl": {
                      "type": [
                        "string"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "personName",
                    "color",
                    "icsUrl"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
    "/api/calendar/feeds/{id}": {
      "delete": {
        "operationId": "CalendarController_removeFeed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": [
                "string"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "object"
                  ],
                  "properties": {
                    "ok": {
                      "type": [
                        "boolean"
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
    "/api/calendar/events": {
      "get": {
        "operationId": "CalendarController_getEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array"
                  ],
                  "items": {
                    "type": [
                      "object"
                    ],
                    "properties": {
                      "uid": {
                        "type": [
                          "string"
                        ]
                      },
                      "summary": {
                        "type": [
                          "string"
                        ]
                      },
                      "start": {
                        "type": [
                          "string"
                        ]
                      },
                      "end": {
                        "type": [
                          "string"
                        ]
                      },
                      "allDay": {
                        "type": [
                          "boolean"
                        ]
                      },
                      "location": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "personName": {
                        "type": [
                          "string"
                        ]
                      },
                      "color": {
                        "type": [
                          "string"
                        ]
                      }
                    },
                    "required": [
                      "uid",
                      "summary",
                      "start",
                      "end",
                      "allDay",
                      "location",
                      "personName",
                      "color"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartFridge API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}